feat: Verbose mode

+++ plugins/kubectl/env-delete.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-
 : "${taito_cli_path:?}"
 : "${taito_plugin_path:?}"
 : "${taito_project:?}"
@@ -15,14 +14,14 @@ echo "Deleting secrets from Kubernetes" && \
 echo "Delete and purge helm release ${taito_project}-${taito_env} (Y/n)?" && \
 read -r confirm && \
 if [[ "${confirm}" =~ ^[Yy]$ ]]; then
-  helm delete --purge "${taito_project}-${taito_env}"
+  (${taito_setv:?}; helm delete --purge "${taito_project}-${taito_env}")
 fi && \
 
 echo "Delete namespace ${taito_namespace} (Y/n)?" && \
 echo "WARNING: Do not delete the namespace if it contains also some other apps" && \
 read -r confirm && \
 if [[ "${confirm}" =~ ^[Yy]$ ]]; then
-  kubectl delete namespace "${taito_namespace}"
+  (${taito_setv:?}; kubectl delete namespace "${taito_namespace}")
 fi && \
 
 # Call next command on command chain
