+++ plugins/kubectl/oper-restart.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-
 : "${taito_cli_path:?}"
 : "${taito_plugin_path:?}"
 : "${taito_namespace:?}"
@@ -7,7 +6,7 @@
 # Change namespace
 "${taito_plugin_path}/util/use-context.sh" && \
 
-kubectl delete pods --all && \
+(${taito_setv:?}; kubectl delete pods --all) && \
 echo "kubectl: TODO rolling restart instead of kill" && \
 
 # Call next command on command chain
