+++ plugins/kubectl/oper-stop.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-
 : "${taito_cli_path:?}"
 : "${taito_plugin_path:?}"
 : "${taito_namespace:?}"
@@ -10,7 +9,7 @@
 "${taito_plugin_path}/util/use-context.sh" && \
 
 echo "TODO scale replicas to 0 instead of deleting" && \
-helm delete --purge "${taito_project}-${taito_env}" && \
+(${taito_setv:?}; helm delete --purge "${taito_project}-${taito_env}") && \
 
 # Call next command on command chain
 "${taito_cli_path}/util/call-next.sh" "${@}"
